Bake in ssh keys for common Git servers. (#107)
We need to populate the known-hosts file to avoid MITM attacks.
Interactively, ssh would ask the user to verify a key, but in a docker
container that's not possible. That's why we fill them beforehand.

If another server is needed, use the ENV variable 'SSH_KNOWN_HOSTS' or
create a new docker image that inherits from this one.